Split out test methods to separate fixtures.
