+++ blueflood-core/src/test/java/com/rackspacecloud/blueflood/service/ScheduleContextTest.java
@@ -388,238 +388,6 @@ public class ScheduleContextTest {
     }
 
     @Test
-    public void testUpdateCreatesActiveDirtyStamp() {
-
-        long now = 1234000L;
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        SlotKey slotkey = SlotKey.of(Granularity.MIN_5, 4, ringShards.get(0));
-
-        // precondition
-        Assert.assertEquals(0, ctx.getScheduledCount());
-        UpdateStamp stamp = ctx.getShardStateManager().getUpdateStamp(slotkey);
-        Assert.assertNull(stamp);
-
-        // when
-        ctx.update(now, ringShards.get(0));
-
-        // then
-        stamp = ctx.getShardStateManager().getUpdateStamp(slotkey);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(now, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-        Assert.assertEquals(0, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testUpdateMarksSlotsDirtyAtAllGranularities() {
-
-        long now = 1234000L;
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ShardStateManager mgr = ctx.getShardStateManager();
-        SlotKey slotkey5 = SlotKey.of(Granularity.MIN_5, 4, ringShards.get(0));
-        SlotKey slotkey20 = SlotKey.of(Granularity.MIN_20, 1, ringShards.get(0));
-        SlotKey slotkey60 = SlotKey.of(Granularity.MIN_60, 0, ringShards.get(0));
-        SlotKey slotkey240 = SlotKey.of(Granularity.MIN_240, 0, ringShards.get(0));
-        SlotKey slotkey1440 = SlotKey.of(Granularity.MIN_1440, 0, ringShards.get(0));
-
-        // precondition
-        Assert.assertEquals(0, ctx.getScheduledCount());
-        Assert.assertNull(mgr.getUpdateStamp(slotkey5));
-        Assert.assertNull(mgr.getUpdateStamp(slotkey20));
-        Assert.assertNull(mgr.getUpdateStamp(slotkey60));
-        Assert.assertNull(mgr.getUpdateStamp(slotkey240));
-        Assert.assertNull(mgr.getUpdateStamp(slotkey1440));
-
-        // when
-        ctx.update(now, ringShards.get(0));
-
-        // then
-        UpdateStamp stamp;
-
-        stamp = mgr.getUpdateStamp(slotkey5);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(now, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(slotkey20);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(now, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(slotkey60);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(now, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(slotkey240);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(now, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(slotkey1440);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(now, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        Assert.assertEquals(0, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testScheduleSlotsOlderThanAddsToScheduledCount() {
-
-        long now = 1234000L;
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(now-2, ringShards.get(0));
-
-        // precondition
-        Assert.assertEquals(0, ctx.getScheduledCount());
-
-        // when
-        ctx.scheduleSlotsOlderThan(1);
-
-        // then
-        Assert.assertEquals(1, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testScheduleSlotsOlderThanSetsHasScheduled() {
-
-        long now = 1234000L;
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(now-2, ringShards.get(0));
-
-        // precondition
-        Assert.assertFalse(ctx.hasScheduled());
-
-        // when
-        ctx.scheduleSlotsOlderThan(1);
-
-        // then
-        Assert.assertTrue(ctx.hasScheduled());
-    }
-
-    @Test
-    public void testGetNextScheduledDecrementsScheduledCount() {
-
-        // given
-        long now = 1234000L;
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        int shard = ringShards.get(0);
-        ctx.update(now-2, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        Assert.assertEquals(1, ctx.getScheduledCount());
-
-        // when
-        SlotKey next = ctx.getNextScheduled();
-
-        // then
-        Assert.assertEquals(0, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testGetNextScheduledIncrementsRunningCount() {
-
-        // given
-        long now = 1234000L;
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        int shard = ringShards.get(0);
-        ctx.update(now-2, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        Assert.assertEquals(0, ctx.getRunningCount());
-
-        // when
-        SlotKey next = ctx.getNextScheduled();
-
-        // then
-        Assert.assertEquals(1, ctx.getRunningCount());
-    }
-
-    @Test
-    public void testGetNextScheduledReturnsTheSmallestGranularity() {
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-        int slot = gran.slot(now);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // when
-        SlotKey next = ctx.getNextScheduled();
-
-        // then
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(slot, next.getSlot());
-        Assert.assertEquals(gran, next.getGranularity());
-    }
-
-    @Test
-    public void testGetNextScheduledChangesStateToRunning() {
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-        int slot = gran.slot(updateTime);
-        Granularity coarserGran = null;
-        try {
-            coarserGran = gran.coarser();
-        } catch (GranularityException e) {
-            Assert.fail("Couldn't get the next coarser granularity");
-        }
-        int coarserSlot = coarserGran.slot(updateTime);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ShardStateManager mgr = ctx.getShardStateManager();
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        UpdateStamp stamp = mgr.getUpdateStamp(SlotKey.of(gran, slot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(SlotKey.of(coarserGran, coarserSlot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        // when
-        SlotKey next = ctx.getNextScheduled();
-
-        // then
-        stamp = mgr.getUpdateStamp(next);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Running, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(SlotKey.of(coarserGran, coarserSlot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-    }
-
-    @Test
     public void testGetNextScheduledSecondTimeReturnsNull() {
 
         // given
@@ -663,126 +431,6 @@ public class ScheduleContextTest {
         // then
         Assert.assertNull(next);
         Assert.assertEquals(0, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testPushBackToScheduledIncrementsScheduledCount() {
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-        int slot = gran.slot(now);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(slot, next.getSlot());
-        Assert.assertEquals(gran, next.getGranularity());
-
-        Assert.assertEquals(0, ctx.getScheduledCount());
-
-        // when
-        ctx.pushBackToScheduled(next, true);
-
-        // then
-        Assert.assertEquals(1, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testPushBackToScheduled_DOES_NOT_DecrementsRunningCount() {
-
-        // TODO: This functionality is probably wrong. It's related to a fairly
-        // rare failure condition so it shouldn't happen often. Nevertheless,
-        // we should fix it so that re-scheduling a slot should pull it out of
-        // the 'running' category, since it's no longer actually running.
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-        int slot = gran.slot(now);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(slot, next.getSlot());
-        Assert.assertEquals(gran, next.getGranularity());
-
-        Assert.assertEquals(1, ctx.getRunningCount());
-
-        // when
-        ctx.pushBackToScheduled(next, true);
-
-        // then
-        Assert.assertEquals(1, ctx.getRunningCount());
-    }
-
-    @Test
-    public void testPushBackToScheduledChangesStateBackToActive() {
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-        int slot = gran.slot(now);
-        Granularity coarserGran = null;
-        try {
-            coarserGran = gran.coarser();
-        } catch (GranularityException e) {
-            Assert.fail("Couldn't get the next coarser granularity");
-        }
-        int coarserSlot = coarserGran.slot(updateTime);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ShardStateManager mgr = ctx.getShardStateManager();
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(slot, next.getSlot());
-        Assert.assertEquals(gran, next.getGranularity());
-
-        UpdateStamp stamp = mgr.getUpdateStamp(SlotKey.of(gran, slot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Running, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(SlotKey.of(coarserGran, coarserSlot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        // when
-        ctx.pushBackToScheduled(next, true);
-
-        // then
-        stamp = mgr.getUpdateStamp(next);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(SlotKey.of(coarserGran, coarserSlot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
     }
 
     @Test
@@ -898,185 +546,4 @@ public class ScheduleContextTest {
         // an NPE is thrown
     }
 
-    @Test
-    public void testPushBackToScheduledRescheduleImmediately() {
-
-        // given
-        long now = 1234000L;
-        long fiveMinutes = 5 * 60 * 1000;
-        long updateTime1 = now - 2;
-        long updateTime2 = now - fiveMinutes - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(updateTime1, shard);
-        ctx.update(updateTime2, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        Assert.assertEquals(2, ctx.getScheduledCount());
-
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(gran, next.getGranularity());
-        int nextSlot = next.getSlot();
-
-        Assert.assertEquals(1, ctx.getScheduledCount());
-
-        // when
-        ctx.pushBackToScheduled(next, true);
-
-        // then
-        Assert.assertEquals(2, ctx.getScheduledCount());
-        next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(nextSlot, next.getSlot());
-        Assert.assertEquals(gran, next.getGranularity());
-        Assert.assertEquals(1, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testPushBackToScheduledDontRescheduleImmediately() {
-
-        // given
-        long now = 1234000L;
-        long fiveMinutes = 5 * 60 * 1000;
-        long updateTime1 = now - 2;
-        long updateTime2 = now - fiveMinutes - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-        int slot1 = gran.slot(now);
-        int slot2 = gran.slot(now - fiveMinutes);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(updateTime1, shard);
-        ctx.update(updateTime2, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        Assert.assertEquals(2, ctx.getScheduledCount());
-
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(gran, next.getGranularity());
-        int nextSlot = next.getSlot();
-        int otherSlot = (nextSlot == slot1 ? slot2 : slot1);
-
-        Assert.assertEquals(1, ctx.getScheduledCount());
-
-        // when
-        ctx.pushBackToScheduled(next, false);
-
-        // then
-        Assert.assertEquals(2, ctx.getScheduledCount());
-        next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(otherSlot, next.getSlot());
-        Assert.assertEquals(gran, next.getGranularity());
-        Assert.assertEquals(1, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testClearFromRunningDecrementsRunningCount() {
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(1, ctx.getRunningCount());
-
-        // when
-        ctx.clearFromRunning(next);
-
-        // then
-        Assert.assertEquals(0, ctx.getRunningCount());
-    }
-
-    @Test
-    public void testClearFromRunningDoesntChangeScheduledCount() {
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(0, ctx.getScheduledCount());
-
-        // when
-        ctx.clearFromRunning(next);
-
-        // then
-        Assert.assertEquals(0, ctx.getScheduledCount());
-    }
-
-    @Test
-    public void testClearFromRunningChangesStateToRolled() {
-
-        // given
-        long now = 1234000L;
-        long updateTime = now - 2;
-        int shard = ringShards.get(0);
-        Granularity gran = Granularity.MIN_5;
-        int slot = gran.slot(now);
-        Granularity coarserGran = null;
-        try {
-            coarserGran = gran.coarser();
-        } catch (GranularityException e) {
-            Assert.fail("Couldn't get the next coarser granularity");
-        }
-        int coarserSlot = coarserGran.slot(updateTime);
-
-        ScheduleContext ctx = new ScheduleContext(now, ringShards);
-        ShardStateManager mgr = ctx.getShardStateManager();
-        ctx.update(updateTime, shard);
-        ctx.scheduleSlotsOlderThan(1);
-
-        // precondition
-        SlotKey next = ctx.getNextScheduled();
-        Assert.assertEquals(shard, next.getShard());
-        Assert.assertEquals(slot, next.getSlot());
-        Assert.assertEquals(gran, next.getGranularity());
-
-        UpdateStamp stamp = mgr.getUpdateStamp(SlotKey.of(gran, slot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Running, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(SlotKey.of(coarserGran, coarserSlot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        // when
-        ctx.clearFromRunning(next);
-
-        // then
-        stamp = mgr.getUpdateStamp(next);
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Rolled, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-
-        stamp = mgr.getUpdateStamp(SlotKey.of(coarserGran, coarserSlot, shard));
-        Assert.assertNotNull(stamp);
-        Assert.assertEquals(UpdateStamp.State.Active, stamp.getState());
-        Assert.assertEquals(updateTime, stamp.getTimestamp());
-        Assert.assertTrue(stamp.isDirty());
-    }
 }
